requirements.txt
Flask==2.3.3
PyPDF2==3.0.1
python-dateutil==2.8.2
